pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-bots"
version = "0.1.0"
description = "Production-grade AI bots and automation (RAG, chat, actions)."
readme = "readme.md"
authors = [{ name = "AI Tool" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "structlog>=24.1",
  "langchain>=0.2.0",
  "langchain-openai>=0.1.21",
  "langchain-text-splitters>=0.2.0",
  "langchain-community>=0.2.0",
  "openai>=1.40.0",
  "chromadb>=0.5.0",
  "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "mypy>=1.10",
  "ruff>=0.6.2",
  "black>=24.8",
  "pre-commit>=3.8",
  "httpx>=0.27",
]

[project.scripts]
ai-bots = "ai_bots.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
pretty = true

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
